Fix various issues with batch source
